+++ src/main/java/com/wanted/user/validation/annotation/PasswordInfoSameValid.java
@@ -1,6 +1,6 @@
-package com.wanted.user.annotation;
+package com.wanted.user.validation.annotation;
 
-import com.wanted.user.annotation.validator.CommonPasswordValidator;
+import com.wanted.user.validation.validator.PasswordInfoSameValidator;
 import jakarta.validation.Constraint;
 import jakarta.validation.Payload;
 import java.lang.annotation.ElementType;
@@ -8,11 +8,11 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target({ElementType.FIELD})
+@Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = CommonPasswordValidator.class)
-public @interface CommonPasswordValid {
-    String message() default "해당 비밀 번호는 사용할 수 없 습니다.";
+@Constraint(validatedBy = PasswordInfoSameValidator.class)
+public @interface PasswordInfoSameValid {
+    String message() default "다른 정보들과 유사한 비밀번호를 사용할 수 없습니다.";
     Class<?>[] groups() default {};
     Class<? extends Payload>[] payload() default {};
 }
